updated at 15/2/2026

+++ app/Http/Controllers/ProfileController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\ProfileUpdateRequest;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\View\View;
 
@@ -40,21 +41,31 @@ class ProfileController extends Controller
     /**
      * Delete the user's account.
      */
-    public function destroy(Request $request): RedirectResponse
-    {
-        $request->validateWithBag('userDeletion', [
-            'password' => ['required', 'current_password'],
-        ]);
+public function destroy(Request $request): RedirectResponse
+{
+    
+    $request->validate( [
+        'passwordDelete' => ['required'],
+    ],[
+        'passwordDelete.required'=>'يجب ادخال كلمه المرور الحاليه',
+    ]);
 
-        $user = $request->user();
-
-        Auth::logout();
-
-        $user->delete();
-
-        $request->session()->invalidate();
-        $request->session()->regenerateToken();
-
-        return Redirect::to('/');
+    //  التحقق من صحة كلمة المرور
+    if (! Hash::check($request->passwordDelete, $request->user()->password)) {
+        
+        return back()->withErrors(['passwordDelete' => 'كلمة المرور غير صحيحه']);
     }
+
+    
+    $user = $request->user();
+
+    Auth::logout();
+
+    $user->delete();
+
+    $request->session()->invalidate();
+    $request->session()->regenerateToken();
+
+    return Redirect::to('/');
+}
 }
